+++ src/moria2.cpp
@@ -298,11 +298,11 @@ void playerFindInitialize(int direction) {
 
     // We must erase the player symbol '@' here, because sub3_move_light()
     // does not erase the previous location of the player when in find mode
-    // and when run_print_self is false.  The player symbol is not draw at all
+    // and when `run_print_self` is false.  The player symbol is not draw at all
     // in this case while moving, so the only problem is on the first turn
     // of find mode, when the initial position of the character must be erased.
     // Hence we must do the erasure here.
-    if (!temporary_light_only && !run_print_self) {
+    if (!temporary_light_only && !config.run_print_self) {
         putChar(caveGetTileSymbol(char_row, char_col), char_row, char_col);
     }
 
@@ -345,7 +345,7 @@ static bool areaAffectStopLookingAtSquares(int i, int dir, int newDir, int y, in
         if (c_ptr->tptr != 0) {
             int tileID = treasure_list[c_ptr->tptr].tval;
 
-            if (tileID != TV_INVIS_TRAP && tileID != TV_SECRET_DOOR && (tileID != TV_OPEN_DOOR || !run_ignore_doors)) {
+            if (tileID != TV_INVIS_TRAP && tileID != TV_SECRET_DOOR && (tileID != TV_OPEN_DOOR || !config.run_ignore_doors)) {
                 playerEndRunning();
                 return true;
             }
@@ -452,7 +452,7 @@ void playerAreaAffect(int direction, int y, int x) {
 
     // choose a direction.
 
-    if (option2 == 0 || (run_examine_corners && !run_cut_corners)) {
+    if (option2 == 0 || (config.run_examine_corners && !config.run_cut_corners)) {
         // There is only one option, or if two, then we always examine
         // potential corners and never cur known corners, so you step
         // into the straight option.
@@ -478,7 +478,7 @@ void playerAreaAffect(int direction, int y, int x) {
     if (!playerCanSeeDungeonWall(option, row, col) || !playerCanSeeDungeonWall(check_dir, row, col)) {
         // Don't see that it is closed off.  This could be a
         // potential corner or an intersection.
-        if (run_examine_corners && playerSeeNothing(option, row, col) && playerSeeNothing(option2, row, col)) {
+        if (config.run_examine_corners && playerSeeNothing(option, row, col) && playerSeeNothing(option2, row, col)) {
             // Can not see anything ahead and in the direction we are
             // turning, assume that it is a potential corner.
             find_direction = option;
@@ -487,7 +487,7 @@ void playerAreaAffect(int direction, int y, int x) {
             // STOP: we are next to an intersection or a room
             playerEndRunning();
         }
-    } else if (run_cut_corners) {
+    } else if (config.run_cut_corners) {
         // This corner is seen to be enclosed; we cut the corner.
         find_direction = option2;
         find_prevdir = option2;